+++ src/misc/cstring.h
@@ -31,15 +31,15 @@ public:
     ~CString();
 
     uint8_t* data();
-    uint16_t size() const;
+    size_t size() const;
 
-    static QString toString(const uint8_t* cMessage, const uint16_t cMessageSize);
-    static uint16_t fromString(const QString& message, uint8_t* cMessage);
+    static QString toString(const uint8_t* cMessage, const size_t cMessageSize);
+    static size_t fromString(const QString& message, uint8_t* cMessage);
 
 private:
     const static int MAX_SIZE_OF_UTF8_ENCODED_CHARACTER = 4;
 
     uint8_t* cString;
-    uint16_t cStringSize;
+    size_t cStringSize;
 };
 #endif // CSTRING_H
